Removing AVFoundation dependency because it clashes with ffmpeg.
AVFoundation is only used with libsourcey to enable listing of internal
video devices (webcams) etc as far as I can see. There is a replacement
function but it might not work. Anyway, I will not be using it so I was
fine tossing it. If you need it, make sure it will work on MacOS using
the replacement function GetQTKitVideoDevices
(devicemanager_mac.mm:146).